requirements.txt
aiogram==3.10.0
aiohttp==3.9.5
supabase==2.16.0
aiocryptopay==0.4.8
gunicorn==22.0.0
google-generativeai==0.7.2
Pillow==10.4.0
